better file structure + update includepath
