+++ firmware/arduino/hal/I2C/I2C-Arduino.hpp
@@ -1,22 +1,24 @@
 #pragma once
 
-#include <HardwareSerial.h>
-#include "UART-Interface.hpp"
+#include "Wire.h"
+#include "interfaces/I2C-Interface.hpp"
 
-
-class UART_Arduino : public UART_Interface
+class I2C_Arduino : public I2C_Interface
 {
 private:
-    HardwareSerial& _serial;
+    TwoWire& _wire;
 public:
-    UART_Arduino(HardwareSerial& serial);
+    I2C_Arduino(TwoWire& wire);
 
-    bool begin(uint32_t baud) override;
+    bool begin(void) override;
     void setTimeout(uint32_t timeout) override;
     uint32_t getTimeout(void) const override;
+    bool beginTransmission(uint8_t I2C_addr) override;
+    bool endTransmission(void) override;
     bool write(uint8_t byte) override;
     size_t writeBytes(const uint8_t *bytes, size_t len) override;
     size_t available(void) override;
+    size_t requestFrom(uint8_t I2C_addr, size_t amount) override;
     int16_t read(void) override;
     size_t readBytes(uint8_t *buffer, size_t len) override;
